+++ libs/components/angular-tree-component/projects/angular-tree-component/src/public-api.ts
@@ -1,0 +1,6 @@
+export * from './modules/angular-tree/angular-tree.module';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of the public API.
+export { SkyAngularTreeWrapperComponent as λ1 } from './modules/angular-tree/angular-tree-wrapper.component';
+export { SkyAngularTreeNodeComponent as λ2 } from './modules/angular-tree/angular-tree-node.component';