+++ libs/components/angular-tree-component/src/app/public/public_api.ts
@@ -1,1 +1,0 @@
-export * from './modules/angular-tree/angular-tree.module';